fix: BMP V14 — u32 posting prefix sums (u16 overflowed on 256-doc blocks, prod SIGSEGV) (#31)
* fix(wasm): allow dead_code on native-only observe stubs and BmpIndex raw-blob accessors

Their only callers (segment::reorder, native directory code) are cfg'd off
the wasm build, so the wasm CI job (-D warnings) failed on dead_code.

* fix: BMP V14 — u32 posting prefix sums (u16 overflowed on 256-doc blocks, prod SIGSEGV)

A block with more than 65,535 postings (bmp_block_size 256 x ~300-dim docs)
silently wrapped the u16 posting prefix sums at build time; the reader's
end - start then underflowed into a multi-exabyte posting slice and the BP
forward-index build (which walks every term of every block) read off the
end of the mmap — SIGSEGV in optimizer-bp threads, CrashLoopBackOff.
Verified against prod: same faulting IP across all crashes, symbolized to
the fwd-build fold; the affected segment averages 21.5k postings/block.

- Format V14 (magic 0x34504D42): per-block num_terms and posting prefix
  sums are u32 (were u16). No compatibility shim — readers loudly refuse
  V13 and older; segments must be rebuilt.
- Reader hardening: block_term_postings refuses non-monotonic prefix sums
  (debug_assert + empty slice) instead of fabricating a wild slice.
- Regression: test_bmp_block_posting_overflow_256 builds a 77k-posting
  block, checks exact needle search past the old wrap point and a clean
  BP forward-index build (the exact crash path). Reproduced first: old
  code panics with "attempt to add with overflow" at the u16 accumulation.

+++ hermes-core/src/query/bmp.rs
@@ -683,7 +683,7 @@ fn zero_touched_acc(acc: &mut [u32], touched: &[u64; 4]) {
 #[inline(always)]
 #[allow(clippy::too_many_arguments)]
 fn score_block_bsearch_int(
-    num_terms: u16,
+    num_terms: u32,
     dim_ptr: *const u8,
     ps_ptr: *const u8,
     post_ptr: *const u8,
